Add the option to pass the authurl as a command line option. Also only retreive and save story if it doesn't already exist on disk

+++ extractAuthor/authscraper.py
@@ -8,24 +8,30 @@ import requests
 from bs4 import BeautifulSoup
 import time
 
+# GEL - pass url via command line and skip existing files
+import sys
+import os
+if sys.argv[1]:
+    authurl=sys.argv[1];
+else:
 
-# In[25]:
+    # In[25]:
 
 
-authurl = "https://www.literotica.com/stories/memberpage.php?uid=1253141&page=submissions"
+    authurl = "https://www.literotica.com/stories/memberpage.php?uid=1253141&page=submissions"
 
 
-# In[26]:
+    # In[26]:
 
 
-authurl = input ("Type literotica author story submissions url \nurl example - https://www.literotica.com/stories/memberpage.php?uid=1253141&page=submissions \nType URL:")
+    authurl = input ("Type literotica author story submissions url \nurl example - https://www.literotica.com/stories/memberpage.php?uid=1253141&page=submissions \nType URL:")
 
 
 # In[27]:
 
 
 
-try:    
+try:
     authpagehtml = requests.get(authurl,headers={"User-Agent" : "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36"})
     print('auth found')
     authsoup = BeautifulSoup(authpagehtml.content, "lxml")
@@ -45,25 +51,25 @@ try:
 except:
     import traceback
     traceback.print_exc()
-    
+
 try:
     for pagedesc in authsoup.select('tr[class*="root-story"]'):
         x = pagedesc.find_all("a")
 
         for page in x:
-            lpstr = page.attrs['href']    
+            lpstr = page.attrs['href']
             if "/s/" in lpstr:
                rootstories.append(lpstr)
 except:
     import traceback
     traceback.print_exc()
-    
+
 try:
     for pagedesc in authsoup.select('tr[class*="sl"]'):
         x = pagedesc.find_all("a")
 
         for page in x:
-            lpstr = page.attrs['href']    
+            lpstr = page.attrs['href']
             if "/s/" in lpstr:
                seriesstories.append(lpstr)
 except:
@@ -84,7 +90,7 @@ print( "stories to extract=> ",allstories)
 def extractStory(author,url):
     import os
     os.makedirs(author,exist_ok=True)
-    try:    
+    try:
         storyhtml = requests.get(url,headers={"User-Agent" : "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36"})
         print('story found')
         soup = BeautifulSoup(storyhtml.content, "lxml")
@@ -102,7 +108,7 @@ def extractStory(author,url):
             x = pagedesc.find_all("a")
 
             for page in x:
-                lpstr = page.attrs['href']    
+                lpstr = page.attrs['href']
                 if "page=" in lpstr:
                     lastpage = int(lpstr.split("page=")[1])
     #         print(lastpage)
@@ -114,7 +120,7 @@ def extractStory(author,url):
 
     pages_to_fetch = []
     pages_to_fetch.append(url)
-    for p in range(2, lastpage+1):        
+    for p in range(2, lastpage+1):
         if lastpage!=0:
             tempurl = url + "?page=" + str(p)
             pages_to_fetch.append(tempurl)
@@ -123,7 +129,7 @@ def extractStory(author,url):
 
     def fetchpage(url):
         story = "ERROR!!!"
-        try:       
+        try:
             storyhtml = requests.get(url,headers={"User-Agent" : "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36"})
             print('fetching data from: ', url)
             soup = BeautifulSoup(storyhtml.content, "lxml")
@@ -139,16 +145,21 @@ def extractStory(author,url):
 
 
     import time
-    fullstory = fname.split(".txt")[0]
-    for page in pages_to_fetch:
-        story = fetchpage(page)
-    #     time.sleep(1)
-        fullstory = fullstory + "\n\nSource:" + page + "\n\n" + story + "\n" 
-    
     fullfilename = author +"/" + fname
-    with open(fullfilename, 'w', encoding="utf-8") as f:
-        f.write(fullstory)
-        print("Story exported as :", fullfilename)
+
+    # GEL - check to see if the fullfilename already exists and skip if we already have it
+    if os.path.exists(fullfilename):
+        print (fullfilename+" already exists, skipping \n");
+    else:
+        fullstory = fname.split(".txt")[0]
+        for page in pages_to_fetch:
+            story = fetchpage(page)
+        #     time.sleep(1)
+            fullstory = fullstory + "\n\nSource:" + page + "\n\n" + story + "\n"
+
+        with open(fullfilename, 'w', encoding="utf-8") as f:
+            f.write(fullstory)
+            print("Story exported as :", fullfilename)
 
 
 # In[31]:
